requirements.txt
pandas
numpy
h3
geopandas
shapely
xgboost
scikit-learn
requests
gdelt
python-dotenv
fastapi
uvicorn
psycopg2-binary
sqlalchemy
supabase
google-genai
apscheduler
tqdm
matplotlib
seaborn
jupyter
